Break up methods.js for SourceAccessRequests
