+++ both/api/source-access-requests/server/methods/index.js
@@ -1,0 +1,2 @@
+export { default as approve } from './approve';
+export { default as askForAccess } from './ask-for-access';
